backend/requirements.txt
# Core
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.3.14
langchain-core>=0.3.28
langsmith>=0.2.10

# LLM Providers
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.14.0
